Small refactoring of library directory Increment Bluz Framework requirements
